LP-928 moved tests to correct folder, relative paths for imports, restructured proposed-name.njk
