+++ src/presentation/test/integration/registration/generalPartner/address/enter-general-partner-usual-residential-address.test.ts
@@ -5,7 +5,7 @@ import cyTranslationText from "../../../../../../../locales/cy/translations.json
 
 import app from "../../../app";
 
-import * as config from "config";
+import * as config from "../../../../../../config";
 import {
   CONFIRM_GENERAL_PARTNER_USUAL_RESIDENTIAL_ADDRESS_URL,
   ENTER_GENERAL_PARTNER_USUAL_RESIDENTIAL_ADDRESS_URL,
