+++ src/presentation/test/integration/transition/address/generalPartner/choose-general-partner-usual-residential-address.test.ts
@@ -1,15 +1,15 @@
 import request from "supertest";
-import enTranslationText from "../../../../../../locales/en/translations.json";
-import cyTranslationText from "../../../../../../locales/cy/translations.json";
-import app from "../../app";
+import enTranslationText from "../../../../../../../locales/en/translations.json";
+import cyTranslationText from "../../../../../../../locales/cy/translations.json";
+import app from "../../../app";
 import {
   CHOOSE_GENERAL_PARTNER_USUAL_RESIDENTIAL_ADDRESS_URL,
   CONFIRM_GENERAL_PARTNER_USUAL_RESIDENTIAL_ADDRESS_URL
-} from "presentation/controller/addressLookUp/url/transition";
-import { getUrl, setLocalesEnabled, testTranslations } from "../../../utils";
-import { appDevDependencies } from "config/dev-dependencies";
-import * as config from "config";
-import AddressPageType from "presentation/controller/addressLookUp/PageType";
+} from "../../../../../../presentation/controller/addressLookUp/url/transition";
+import { getUrl, setLocalesEnabled, testTranslations } from "../../../../utils";
+import { appDevDependencies } from "../../../../../../config/dev-dependencies";
+import * as config from "../../../../../../config";
+import AddressPageType from "../../../../../../presentation/controller/addressLookUp/PageType";
 
 describe("Choose usual residential address of the general partner page", () => {
   beforeEach(() => {
