+++ src/presentation/test/integration/transition/address/generalPartner/territory-choice-general-partner-usual-residential-address.test.ts
@@ -1,17 +1,17 @@
 import request from "supertest";
-import enTranslationText from "../../../../../../locales/en/translations.json";
-import cyTranslationText from "../../../../../../locales/cy/translations.json";
-import app from "../../app";
-import { appDevDependencies } from "../../../../../config/dev-dependencies";
-import { getUrl, setLocalesEnabled, testTranslations } from "../../../utils";
+import enTranslationText from "../../../../../../../locales/en/translations.json";
+import cyTranslationText from "../../../../../../../locales/cy/translations.json";
+import app from "../../../app";
+import { appDevDependencies } from "../../../../../../config/dev-dependencies";
+import { getUrl, setLocalesEnabled, testTranslations } from "../../../../utils";
 import {
   ENTER_GENERAL_PARTNER_USUAL_RESIDENTIAL_ADDRESS_URL,
   TERRITORY_CHOICE_GENERAL_PARTNER_USUAL_RESIDENTIAL_ADDRESS_URL,
   POSTCODE_GENERAL_PARTNER_USUAL_RESIDENTIAL_ADDRESS_URL
-} from "../../../../controller/addressLookUp/url/transition";
-import AddressPageType from "../../../../controller/addressLookUp/PageType";
-import GeneralPartnerBuilder, { generalPartnerPerson } from "../../../builder/GeneralPartnerBuilder";
-import { APPLICATION_CACHE_KEY } from "../../../../../config/constants";
+} from "../../../../../controller/addressLookUp/url/transition";
+import AddressPageType from "../../../../../controller/addressLookUp/PageType";
+import GeneralPartnerBuilder, { generalPartnerPerson } from "../../../../builder/GeneralPartnerBuilder";
+import { APPLICATION_CACHE_KEY } from "../../../../../../config/constants";
 
 describe("General Partner Usual Residential Address Territory Choice", () => {
   const URL = getUrl(TERRITORY_CHOICE_GENERAL_PARTNER_USUAL_RESIDENTIAL_ADDRESS_URL);
